ToDotNot unit test conversion - part 9.
