+++ src/Json.Schema.ToDotNet.UnitTests/DataModelGeneratorTests.cs
@@ -16,9 +16,15 @@ namespace Microsoft.Json.Schema.ToDotNet.UnitTests
 
         // File names for commonly used input data files.
         private const string SchemaFileName = "Schema.json";
+        private const string HintsFileName = "CodeGenHints.json";
+
+        // File names for commonly used expected result files.
         private const string ExpectedClassFileName = "ExpectedClass.cs";
         private const string ExpectedComparerClassFileName = "ExpectedComparerClass.cs";
-        private const string HintsFileName = "CodeGenHints.json";
+        private const string ExpectedSyntaxInterfaceFileName = "ExpectedSyntaxInterface.cs";
+        private const string ExpectedKindEnumFileName = "ExpectedKindEnum.cs";
+        private const string ExpectedEnumTypeFileName = "ExpectedEnumType.cs";
+        private const string ExpectedRewritingVisitorFileName = "ExpectedRewritingVisitor.cs";
 
         private static readonly string PrimaryOutputFilePath = TestFileSystem.MakeOutputFilePath(TestSettings.RootClassName);
 
@@ -288,10 +294,10 @@ namespace Microsoft.Json.Schema.ToDotNet.UnitTests
 
             string hintsText = TestUtil.ReadTestInputFile(ClassName, HintsFileName);
             string expectedClass = TestUtil.ReadTestInputFile(ClassName, ExpectedClassFileName);
-            string expectedSyntaxInterface = TestUtil.ReadTestInputFile(ClassName, "ExpectedSyntaxInterface.cs");
-            string expectedKindEnum = TestUtil.ReadTestInputFile(ClassName, "ExpectedKindEnum.cs");
-            string expectedRewritingVisitor = TestUtil.ReadTestInputFile(ClassName, "ExpectedRewritingVisitor.cs");
-            string expectedEnumType = TestUtil.ReadTestInputFile(ClassName, "ExpectedEnumType.cs");
+            string expectedSyntaxInterface = TestUtil.ReadTestInputFile(ClassName, ExpectedSyntaxInterfaceFileName);
+            string expectedKindEnum = TestUtil.ReadTestInputFile(ClassName, ExpectedKindEnumFileName);
+            string expectedRewritingVisitor = TestUtil.ReadTestInputFile(ClassName, ExpectedRewritingVisitorFileName);
+            string expectedEnumType = TestUtil.ReadTestInputFile(ClassName, ExpectedEnumTypeFileName);
 
             _settings.GenerateCloningCode = true;
             _settings.HintDictionary = new HintDictionary(hintsText);
@@ -495,132 +501,22 @@ namespace Microsoft.Json.Schema.ToDotNet.UnitTests
             _settings.GenerateEqualityComparers = true;
             var generator = new DataModelGenerator(_settings, _testFileSystem.FileSystem);
 
-            JsonSchema schema = SchemaReader.ReadSchema(
-@"{
-  ""type"": ""object"",
-  ""properties"": {
-    ""arrayOfIntByRef"": {
-      ""type"": ""array"",
-      ""items"": {
-        ""$ref"": ""#/definitions/d""
-      }
-    }
-  },
-  ""definitions"": {
-    ""d"": {
-      ""type"": ""integer"",
-      }
-    }
-  }
-}", TestUtil.TestFilePath);
+            string schemaText = TestUtil.ReadTestInputFile(ClassName, SchemaFileName);
+            JsonSchema schema = SchemaReader.ReadSchema(schemaText, TestUtil.TestFilePath);
 
-            const string ExpectedClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class C
-    {
-        public static IEqualityComparer<C> ValueComparer => CEqualityComparer.Instance;
-
-        public bool ValueEquals(C other) => ValueComparer.Equals(this, other);
-        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
-
-        [DataMember(Name = ""arrayOfIntByRef"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<int> ArrayOfIntByRef { get; set; }
-    }
-}";
-            const string ExpectedComparerClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-
-namespace N
-{
-    /// <summary>
-    /// Defines methods to support the comparison of objects of type C for equality.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    internal sealed class CEqualityComparer : IEqualityComparer<C>
-    {
-        internal static readonly CEqualityComparer Instance = new CEqualityComparer();
-
-        public bool Equals(C left, C right)
-        {
-            if (ReferenceEquals(left, right))
-            {
-                return true;
-            }
-
-            if (ReferenceEquals(left, null) || ReferenceEquals(right, null))
-            {
-                return false;
-            }
-
-            if (!object.ReferenceEquals(left.ArrayOfIntByRef, right.ArrayOfIntByRef))
-            {
-                if (left.ArrayOfIntByRef == null || right.ArrayOfIntByRef == null)
-                {
-                    return false;
-                }
-
-                if (left.ArrayOfIntByRef.Count != right.ArrayOfIntByRef.Count)
-                {
-                    return false;
-                }
-
-                for (int index_0 = 0; index_0 < left.ArrayOfIntByRef.Count; ++index_0)
-                {
-                    if (left.ArrayOfIntByRef[index_0] != right.ArrayOfIntByRef[index_0])
-                    {
-                        return false;
-                    }
-                }
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(C obj)
-        {
-            if (ReferenceEquals(obj, null))
-            {
-                return 0;
-            }
-
-            int result = 17;
-            unchecked
-            {
-                if (obj.ArrayOfIntByRef != null)
-                {
-                    foreach (var value_0 in obj.ArrayOfIntByRef)
-                    {
-                        result = result * 31;
-                        result = (result * 31) + value_0.GetHashCode();
-                    }
-                }
-            }
-
-            return result;
-        }
-    }
-}";
+            string expectedClass = TestUtil.ReadTestInputFile(ClassName, ExpectedClassFileName);
+            string expectedComparerClass = TestUtil.ReadTestInputFile(ClassName, ExpectedComparerClassFileName);
 
             string actual = generator.Generate(schema);
 
-            TestUtil.WriteTestResultFiles(ExpectedClass, actual, nameof(GeneratesArrayOfPrimitiveTypeByReference));
+            TestUtil.WriteTestResultFiles(expectedClass, actual, nameof(GeneratesArrayOfPrimitiveTypeByReference));
 
             var expectedContentsDictionary = new Dictionary<string, ExpectedContents>
             {
                 [_settings.RootClassName] = new ExpectedContents
                 {
-                    ClassContents = ExpectedClass,
-                    ComparerClassContents = ExpectedComparerClass
+                    ClassContents = expectedClass,
+                    ComparerClassContents = expectedComparerClass
                 },
                 ["D"] = new ExpectedContents()
             };
@@ -631,319 +527,14 @@ namespace N
         [Fact(DisplayName = "DataModelGenerator generates array of uri-formatted strings")]
         public void GeneratesArrayOfUriFormattedStrings()
         {
-            JsonSchema schema = SchemaReader.ReadSchema(
-@"{
-  ""type"": ""object"",
-  ""properties"": {
-    ""uriFormattedStrings"": {
-      ""type"": ""array"",
-      ""items"": {
-        ""type"": ""string"",
-        ""format"": ""uri""
-      }
-    }
-  }
-}", TestUtil.TestFilePath);
+            string schemaText = TestUtil.ReadTestInputFile(ClassName, SchemaFileName);
+            JsonSchema schema = SchemaReader.ReadSchema(schemaText, TestUtil.TestFilePath);
 
-            const string ExpectedClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class C : ISNode
-    {
-        public static IEqualityComparer<C> ValueComparer => CEqualityComparer.Instance;
-
-        public bool ValueEquals(C other) => ValueComparer.Equals(this, other);
-        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
-
-        /// <summary>
-        /// Gets a value indicating the type of object implementing <see cref=""ISNode"" />.
-        /// </summary>
-        public SNodeKind SNodeKind
-        {
-            get
-            {
-                return SNodeKind.C;
-            }
-        }
-
-        [DataMember(Name = ""uriFormattedStrings"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<Uri> UriFormattedStrings { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class.
-        /// </summary>
-        public C()
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class from the supplied values.
-        /// </summary>
-        /// <param name=""uriFormattedStrings"">
-        /// An initialization value for the <see cref=""P:UriFormattedStrings"" /> property.
-        /// </param>
-        public C(IEnumerable<Uri> uriFormattedStrings)
-        {
-            Init(uriFormattedStrings);
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref=""C"" /> class from the specified instance.
-        /// </summary>
-        /// <param name=""other"">
-        /// The instance from which the new instance is to be initialized.
-        /// </param>
-        /// <exception cref=""ArgumentNullException"">
-        /// Thrown if <paramref name=""other"" /> is null.
-        /// </exception>
-        public C(C other)
-        {
-            if (other == null)
-            {
-                throw new ArgumentNullException(nameof(other));
-            }
-
-            Init(other.UriFormattedStrings);
-        }
-
-        ISNode ISNode.DeepClone()
-        {
-            return DeepCloneCore();
-        }
-
-        /// <summary>
-        /// Creates a deep copy of this instance.
-        /// </summary>
-        public C DeepClone()
-        {
-            return (C)DeepCloneCore();
-        }
-
-        private ISNode DeepCloneCore()
-        {
-            return new C(this);
-        }
-
-        private void Init(IEnumerable<Uri> uriFormattedStrings)
-        {
-            if (uriFormattedStrings != null)
-            {
-                var destination_0 = new List<Uri>();
-                foreach (var value_0 in uriFormattedStrings)
-                {
-                    destination_0.Add(value_0);
-                }
-
-                UriFormattedStrings = destination_0;
-            }
-        }
-    }
-}";
-
-            const string ExpectedComparerClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-
-namespace N
-{
-    /// <summary>
-    /// Defines methods to support the comparison of objects of type C for equality.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    internal sealed class CEqualityComparer : IEqualityComparer<C>
-    {
-        internal static readonly CEqualityComparer Instance = new CEqualityComparer();
-
-        public bool Equals(C left, C right)
-        {
-            if (ReferenceEquals(left, right))
-            {
-                return true;
-            }
-
-            if (ReferenceEquals(left, null) || ReferenceEquals(right, null))
-            {
-                return false;
-            }
-
-            if (!object.ReferenceEquals(left.UriFormattedStrings, right.UriFormattedStrings))
-            {
-                if (left.UriFormattedStrings == null || right.UriFormattedStrings == null)
-                {
-                    return false;
-                }
-
-                if (left.UriFormattedStrings.Count != right.UriFormattedStrings.Count)
-                {
-                    return false;
-                }
-
-                for (int index_0 = 0; index_0 < left.UriFormattedStrings.Count; ++index_0)
-                {
-                    if (left.UriFormattedStrings[index_0] != right.UriFormattedStrings[index_0])
-                    {
-                        return false;
-                    }
-                }
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(C obj)
-        {
-            if (ReferenceEquals(obj, null))
-            {
-                return 0;
-            }
-
-            int result = 17;
-            unchecked
-            {
-                if (obj.UriFormattedStrings != null)
-                {
-                    foreach (var value_0 in obj.UriFormattedStrings)
-                    {
-                        result = result * 31;
-                        if (value_0 != null)
-                        {
-                            result = (result * 31) + value_0.GetHashCode();
-                        }
-                    }
-                }
-            }
-
-            return result;
-        }
-    }
-}";
-
-            const string ExpectedSyntaxInterface =
-@"using System.CodeDom.Compiler;
-
-namespace N
-{
-    /// <summary>
-    /// An interface for all types generated from the S schema.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public interface ISNode
-    {
-        /// <summary>
-        /// Gets a value indicating the type of object implementing <see cref=""ISNode"" />.
-        /// </summary>
-        SNodeKind SNodeKind { get; }
-
-        /// <summary>
-        /// Makes a deep copy of this instance.
-        /// </summary>
-        ISNode DeepClone();
-    }
-}";
-
-            const string ExpectedKindEnum =
-@"using System.CodeDom.Compiler;
-
-namespace N
-{
-    /// <summary>
-    /// A set of values for all the types that implement <see cref=""ISNode"" />.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public enum SNodeKind
-    {
-        /// <summary>
-        /// An uninitialized kind.
-        /// </summary>
-        None,
-        /// <summary>
-        /// A value indicating that the <see cref=""ISNode"" /> object is of type <see cref=""C"" />.
-        /// </summary>
-        C
-    }
-}";
-
-            const string ExpectedRewritingVisitor =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace N
-{
-    /// <summary>
-    /// Rewriting visitor for the S object model.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public abstract class SRewritingVisitor
-    {
-        /// <summary>
-        /// Starts a rewriting visit of a node in the S object model.
-        /// </summary>
-        /// <param name=""node"">
-        /// The node to rewrite.
-        /// </param>
-        /// <returns>
-        /// A rewritten instance of the node.
-        /// </returns>
-        public virtual object Visit(ISNode node)
-        {
-            return this.VisitActual(node);
-        }
-
-        /// <summary>
-        /// Visits and rewrites a node in the S object model.
-        /// </summary>
-        /// <param name=""node"">
-        /// The node to rewrite.
-        /// </param>
-        /// <returns>
-        /// A rewritten instance of the node.
-        /// </returns>
-        public virtual object VisitActual(ISNode node)
-        {
-            if (node == null)
-            {
-                throw new ArgumentNullException(""node"");
-            }
-
-            switch (node.SNodeKind)
-            {
-                case SNodeKind.C:
-                    return VisitC((C)node);
-                default:
-                    return node;
-            }
-        }
-
-        private T VisitNullChecked<T>(T node) where T : class, ISNode
-        {
-            if (node == null)
-            {
-                return null;
-            }
-
-            return (T)Visit(node);
-        }
-
-        public virtual C VisitC(C node)
-        {
-            if (node != null)
-            {
-            }
-
-            return node;
-        }
-    }
-}";
+            string expectedClass = TestUtil.ReadTestInputFile(ClassName, ExpectedClassFileName);
+            string expectedComparerClass = TestUtil.ReadTestInputFile(ClassName, ExpectedComparerClassFileName);
+            string expectedSyntaxInterface = TestUtil.ReadTestInputFile(ClassName, ExpectedSyntaxInterfaceFileName);
+            string expectedKindEnum = TestUtil.ReadTestInputFile(ClassName, ExpectedKindEnumFileName);
+            string expectedRewritingVisitor = TestUtil.ReadTestInputFile(ClassName, ExpectedRewritingVisitorFileName);
 
             _settings.GenerateEqualityComparers = true;
             _settings.GenerateCloningCode = true;
@@ -968,170 +559,34 @@ namespace N
             _testFileSystem.Files.Count.Should().Be(expectedOutputFiles.Count);
             _testFileSystem.Files.Should().OnlyContain(path => expectedOutputFiles.Contains(path));
 
-            _testFileSystem[PrimaryOutputFilePath].Should().Be(ExpectedClass);
-            _testFileSystem[equalityComparerOutputFilePath].Should().Be(ExpectedComparerClass);
-            _testFileSystem[syntaxInterfacePath].Should().Be(ExpectedSyntaxInterface);
-            _testFileSystem[kindEnumPath].Should().Be(ExpectedKindEnum);
-            _testFileSystem[rewritingVisitorClassPath].Should().Be(ExpectedRewritingVisitor);
+            _testFileSystem[PrimaryOutputFilePath].Should().Be(expectedClass);
+            _testFileSystem[equalityComparerOutputFilePath].Should().Be(expectedComparerClass);
+            _testFileSystem[syntaxInterfacePath].Should().Be(expectedSyntaxInterface);
+            _testFileSystem[kindEnumPath].Should().Be(expectedKindEnum);
+            _testFileSystem[rewritingVisitorClassPath].Should().Be(expectedRewritingVisitor);
         }
 
         [Fact(DisplayName = "DataModelGenerator generates array of arrays of primitive type")]
         public void GeneratesArrayOfArraysOfPrimitiveType()
         {
-            const string Schema =
-@"{
-  ""type"": ""object"",
-  ""properties"": {
-    ""arrayOfArrayOfInt"": {
-      ""type"": ""array"",
-      ""items"": {
-        ""$ref"": ""#/definitions/itemType""
-      }
-    }
-  },
-  ""definitions"": {
-    ""itemType"": {
-      ""type"": ""array"",
-      ""items"": {
-        ""type"": ""integer""
-      }
-    }
-  }
-}";
+            string schemaText = TestUtil.ReadTestInputFile(ClassName, SchemaFileName);
+            string expectedClass = TestUtil.ReadTestInputFile(ClassName, ExpectedClassFileName);
+            string expectedComparerClass = TestUtil.ReadTestInputFile(ClassName, ExpectedComparerClassFileName);
 
-            const string ExpectedClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-
-namespace N
-{
-    [DataContract]
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    public partial class C
-    {
-        public static IEqualityComparer<C> ValueComparer => CEqualityComparer.Instance;
-
-        public bool ValueEquals(C other) => ValueComparer.Equals(this, other);
-        public int ValueGetHashCode() => ValueComparer.GetHashCode(this);
-
-        [DataMember(Name = ""arrayOfArrayOfInt"", IsRequired = false, EmitDefaultValue = false)]
-        public IList<IList<int>> ArrayOfArrayOfInt { get; set; }
-    }
-}";
-            const string ExpectedComparerClass =
-@"using System;
-using System.CodeDom.Compiler;
-using System.Collections.Generic;
-
-namespace N
-{
-    /// <summary>
-    /// Defines methods to support the comparison of objects of type C for equality.
-    /// </summary>
-    [GeneratedCode(""Microsoft.Json.Schema.ToDotNet"", """ + VersionConstants.FileVersion + @""")]
-    internal sealed class CEqualityComparer : IEqualityComparer<C>
-    {
-        internal static readonly CEqualityComparer Instance = new CEqualityComparer();
-
-        public bool Equals(C left, C right)
-        {
-            if (ReferenceEquals(left, right))
-            {
-                return true;
-            }
-
-            if (ReferenceEquals(left, null) || ReferenceEquals(right, null))
-            {
-                return false;
-            }
-
-            if (!object.ReferenceEquals(left.ArrayOfArrayOfInt, right.ArrayOfArrayOfInt))
-            {
-                if (left.ArrayOfArrayOfInt == null || right.ArrayOfArrayOfInt == null)
-                {
-                    return false;
-                }
-
-                if (left.ArrayOfArrayOfInt.Count != right.ArrayOfArrayOfInt.Count)
-                {
-                    return false;
-                }
-
-                for (int index_0 = 0; index_0 < left.ArrayOfArrayOfInt.Count; ++index_0)
-                {
-                    if (!object.ReferenceEquals(left.ArrayOfArrayOfInt[index_0], right.ArrayOfArrayOfInt[index_0]))
-                    {
-                        if (left.ArrayOfArrayOfInt[index_0] == null || right.ArrayOfArrayOfInt[index_0] == null)
-                        {
-                            return false;
-                        }
-
-                        if (left.ArrayOfArrayOfInt[index_0].Count != right.ArrayOfArrayOfInt[index_0].Count)
-                        {
-                            return false;
-                        }
-
-                        for (int index_1 = 0; index_1 < left.ArrayOfArrayOfInt[index_0].Count; ++index_1)
-                        {
-                            if (left.ArrayOfArrayOfInt[index_0][index_1] != right.ArrayOfArrayOfInt[index_0][index_1])
-                            {
-                                return false;
-                            }
-                        }
-                    }
-                }
-            }
-
-            return true;
-        }
-
-        public int GetHashCode(C obj)
-        {
-            if (ReferenceEquals(obj, null))
-            {
-                return 0;
-            }
-
-            int result = 17;
-            unchecked
-            {
-                if (obj.ArrayOfArrayOfInt != null)
-                {
-                    foreach (var value_0 in obj.ArrayOfArrayOfInt)
-                    {
-                        result = result * 31;
-                        if (value_0 != null)
-                        {
-                            foreach (var value_1 in value_0)
-                            {
-                                result = result * 31;
-                                result = (result * 31) + value_1.GetHashCode();
-                            }
-                        }
-                    }
-                }
-            }
-
-            return result;
-        }
-    }
-}";
             _settings.GenerateEqualityComparers = true;
             var generator = new DataModelGenerator(_settings, _testFileSystem.FileSystem);
-            JsonSchema schema = SchemaReader.ReadSchema(Schema, TestUtil.TestFilePath);
+            JsonSchema schema = SchemaReader.ReadSchema(schemaText, TestUtil.TestFilePath);
 
             string actual = generator.Generate(schema);
 
-            TestUtil.WriteTestResultFiles(ExpectedClass, actual, nameof(GeneratesArrayOfArraysOfPrimitiveType));
+            TestUtil.WriteTestResultFiles(expectedClass, actual, nameof(GeneratesArrayOfArraysOfPrimitiveType));
 
             var expectedContentsDictionary = new Dictionary<string, ExpectedContents>
             {
                 [_settings.RootClassName] = new ExpectedContents
                 {
-                    ClassContents = ExpectedClass,
-                    ComparerClassContents = ExpectedComparerClass
+                    ClassContents = expectedClass,
+                    ComparerClassContents = expectedComparerClass
                 }
             };
