+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesArrayOfPrimitiveTypeByReference/Schema.json
@@ -1,0 +1,17 @@
+{
+  "type": "object",
+  "properties": {
+    "arrayOfIntByRef": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/d"
+      }
+    }
+  },
+  "definitions": {
+    "d": {
+      "type": "integer",
+      }
+    }
+  }
+}
