+++ src/Json.Schema.ToDotNet.UnitTests/TestData/DataModelGeneratorTests/GeneratesArrayOfUriFormattedStrings/Schema.json
@@ -1,0 +1,12 @@
+{
+  "type": "object",
+  "properties": {
+    "uriFormattedStrings": {
+      "type": "array",
+      "items": {
+        "type": "string",
+        "format": "uri"
+      }
+    }
+  }
+}
